fix: simulation: update screen before sleep

+++ lib/Epub/Epub/BookMetadataCache.cpp
@@ -451,7 +451,7 @@ size_t BookMetadataCache::getSpineCumulativeSize(const int index) {
   serialization::readPod(bookFile, hrefLen);
   bookFile.seekCur(hrefLen);
 
-  size_t cumulativeSize = 0;
+  uint32_t cumulativeSize = 0; // simulator patch
   serialization::readPod(bookFile, cumulativeSize);
   return cumulativeSize;
 }
